app/models/model4.py
"""
Model4 (Template)
"""
from PIL import Image

predict = None

try:
    def predict(image: Image.Image):
        return "not_implemented", 0.0

    predict.framework = "tensorflow"
    predict.input_size = 224

    print("Model4 template loaded successfully.")

except Exception as e:
    print(f"[ERROR] Could not initialize Model4: {e}")

    def predict(image: Image.Image):
        return "not_implemented", 0.0

    predict.framework = "tensorflow"
    predict.input_size = 224
