requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pydantic
numpy
torch
torchvision
tensorflow
